Initial commit of Sapphire for FFXIV 3.0 "light rewrite"

+++ deps/datReader/DatCategories/bg/sgb.h
@@ -136,7 +136,7 @@ struct SGB_MODEL_ENTRY : public SGB_GROUP_ENTRY
   std::string modelFileName;
   std::string collisionFileName;
 
-  SGB_MODEL_ENTRY( char* buf, uint32_t offset, SgbGroupEntryType type )
+  SGB_MODEL_ENTRY( char* buf, size_t offset, SgbGroupEntryType type )
   {
     this->type = type;
     header = *reinterpret_cast< SGB_MODEL_HEADER* >( buf + offset );
@@ -258,7 +258,7 @@ struct SGB_FILE
       if( stateCount > 0 )
       {
         stateCount = stateCount;
-        for( int i = 0; i < stateCount; ++i )
+        for( size_t i = 0; i < stateCount; ++i )
         {
           auto state = SGB_STATE_ENTRY( buf + baseOffset + header.statesOffset + 8 + i * sizeof( SGB_STATE_HEADER ) );
           stateEntries.push_back( state );
